demo/requirements.txt
# CloverShield Demo - Python Dependencies
# Core framework
streamlit>=1.28.0

# Data processing
pandas>=1.5.0
numpy>=1.23.0

# Visualization
plotly>=5.14.0

# Model loading and ML dependencies
joblib>=1.2.0
scikit-learn>=1.2.0
xgboost>=1.7.0

# Explainability
shap>=0.44.0

# LLM explanations (optional)
groq>=1.0.0

# Environment variable management
python-dotenv>=1.0.0

# Graph features (for model pipeline)
networkx>=3.0

# Optional: For better performance
scipy>=1.10.0

